Support snapshots and released jars from the same test script

+++ test_genomicsdbjar.sh
@@ -1,13 +1,34 @@
 #!/bin/bash
 
-export GENOMICSDB_VERSION=1.1.1
+####
+# Set the following variables as necessary
+####
+IS_SNAPSHOT=false
+
+if [ ${IS_SNAPSHOT} == true ]; then
+		export GENOMICSDB_VERSION=1.1.2-20191008.130318-1
+		export GENOMICSB_REPOSITORY_VERSION=1.1.2-SNAPSHOT
+		export MAVEN_REPOSITORY=https://oss.sonatype.org/content/repositories/snapshots/org/genomicsdb/genomicsdb/${GENOMICSB_REPOSITORY_VERSION}
+else
+	export GENOMICSDB_VERSION=1.1.1
+	#export MAVEN_REPOSITORY=https://oss.sonatype.org/content/repositories/staging/org/genomicsdb/genomicsdb/${GENOMICSDB_VERSION}
+  export MAVEN_REPOSITORY=https://repo1.maven.org/maven2/org/genomicsdb/genomicsdb/${GENOMICSDB_VERSION}
+fi
+
+####
+# Leave the rest of the script alone
+####
+
 export CLASSPATH=genomicsdb-${GENOMICSDB_VERSION}-allinone.jar:.
-export MAVEN_REPOSITORY=https://oss.sonatype.org/content/repositories/staging/org/genomicsdb/genomicsdb/${GENOMICSDB_VERSION}
 
 curl -O ${MAVEN_REPOSITORY}/genomicsdb-${GENOMICSDB_VERSION}-allinone.jar
 curl -O ${MAVEN_REPOSITORY}/genomicsdb-${GENOMICSDB_VERSION}.jar
 
-bash -x ./run_checks.sh
+./run_checks.sh
+if [[ $? -ne 0 ]]; then
+    echo "run_checks FAILED"
+    exit 1
+fi
 
 rm -fr GenomicsDB
 git clone https://github.com/GenomicsDB/GenomicsDB.git
@@ -20,17 +41,17 @@ echo &&
 
 echo "Starting ETL..." &&
 java TestGenomicsDBImporterWithMergedVCFHeader -L 1:1-100000 -w /tmp/ws -A test0 GenomicsDB/tests/inputs/vcfs/t0.vcf.gz GenomicsDB/tests/inputs/vcfs/t1.vcf.gz GenomicsDB/tests/inputs/vcfs/t2.vcf.gz --vidmap-output /tmp/vid_pb.json --callset-output /tmp/callset_pb.json &&
-echo "Done" &&
+echo "ETL SUCCESS" &&
 echo &&
 
 echo "Starting query..." &&
 java TestGenomicsDB --query -l loader.json query.json &&
-echo "Done" &&
+echo "Query SUCCESS" &&
 echo
 
 if [[ $? -ne 0 ]]; then
-        echo "TestGenomicsDBJAR FAILED"
-        exit 1
+    echo "TestGenomicsDBJAR FAILED"
+    exit 1
 fi
